refactor: consolidate client auth patches into one module
patches/utils.py duplicated the token helpers already present in
patches/client/common.py; the async client used one copy and the sync
client the other. Merge them into patches/client/common.py (imported by
both sync.py and aio.py) and delete patches/utils.py.

With a single home for the helpers, flip their verify parameter to the
native httpx2 type (bool | ssl.SSLContext) and convert the CA path once in
each client __init__, keeping the bool/str form only for azure-core's
connection_verify.
