+++ diracx-client/src/diracx/client/patches/client/common.py
@@ -3,12 +3,15 @@
 from __future__ import annotations
 
 __all__ = [
-    "DiracAuthMixin",
+    "get_openid_configuration",
+    "get_token",
+    "refresh_token",
 ]
 
 import fcntl
 import json
 import os
+import ssl
 from datetime import datetime, timedelta, timezone
 from dataclasses import dataclass
 from enum import Enum
@@ -21,7 +24,6 @@ import jwt
 from azure.core.credentials import AccessToken
 from diracx.core.utils import (
     EXPIRES_GRACE_SECONDS,
-    prepare_verify,
     serialize_credentials,
 )
 from diracx.core.models import TokenResponse
@@ -41,12 +43,12 @@ class TokenResult:
 
 
 def get_openid_configuration(
-    endpoint: str, *, verify: bool | str = True
+    endpoint: str, *, verify: bool | ssl.SSLContext = True
 ) -> dict[str, str]:
     """Get the openid configuration from the .well-known endpoint"""
     response = httpx2.get(
         url=parse.urljoin(endpoint, ".well-known/openid-configuration"),
-        verify=prepare_verify(verify),
+        verify=verify,
     )
     if not response.is_success:
         raise RuntimeError("Cannot fetch any information from the .well-known endpoint")
@@ -58,7 +60,7 @@ def get_token(
     token: AccessToken | None,
     token_endpoint: str,
     client_id: str,
-    verify: bool | str,
+    verify: bool | ssl.SSLContext,
 ) -> AccessToken:
     """Get the access token if available and still valid."""
     # Immediately return the token if it is available and still valid
@@ -121,7 +123,7 @@ def refresh_token(
     client_id: str,
     refresh_token: str,
     *,
-    verify: bool | str = True,
+    verify: bool | ssl.SSLContext = True,
 ) -> TokenResponse:
     """Refresh the access token using the refresh_token flow."""
     response = httpx2.post(
@@ -131,7 +133,7 @@ def refresh_token(
             "grant_type": "refresh_token",
             "refresh_token": refresh_token,
         },
-        verify=prepare_verify(verify),
+        verify=verify,
     )
 
     if response.status_code != 200:
