+++ diracx-client/src/diracx/client/patches/client/sync.py
@@ -6,6 +6,7 @@ __all__ = [
     "Dirac",
 ]
 
+import ssl
 from importlib.metadata import PackageNotFoundError, distribution
 from pathlib import Path
 from typing import Any, Optional
@@ -14,6 +15,7 @@ from azure.core.credentials import AccessToken, TokenCredential
 from azure.core.pipeline import PipelineRequest
 from azure.core.pipeline.policies import BearerTokenCredentialPolicy
 from diracx.core.preferences import DiracxPreferences, get_diracx_preferences
+from diracx.core.utils import prepare_verify
 
 from .common import get_openid_configuration, get_token
 from ..._generated._client import Dirac as _Dirac
@@ -28,7 +30,7 @@ class SyncDiracTokenCredential(TokenCredential):
         token_endpoint: str,
         client_id: str,
         *,
-        verify: bool | str = True,
+        verify: bool | ssl.SSLContext = True,
     ) -> None:
         self.location = location
         self.verify = verify
@@ -104,10 +106,15 @@ class Dirac(_Dirac):
         if verify is True and diracx_preferences.ca_path:
             verify = str(diracx_preferences.ca_path)
         kwargs["connection_verify"] = verify
+        # azure-core wants a bool/path for connection_verify, but httpx2 wants
+        # an SSLContext, so convert the CA path once here for the httpx2 calls.
+        ssl_verify = prepare_verify(verify)
         self._client_id = client_id or "myDIRACClientID"
 
         # Get .well-known configuration
-        openid_configuration = get_openid_configuration(self._endpoint, verify=verify)
+        openid_configuration = get_openid_configuration(
+            self._endpoint, verify=ssl_verify
+        )
 
         try:
             self.client_version = distribution("diracx").version
@@ -130,7 +137,7 @@ class Dirac(_Dirac):
                     location=diracx_preferences.credentials_path,
                     token_endpoint=openid_configuration["token_endpoint"],
                     client_id=self._client_id,
-                    verify=verify,
+                    verify=ssl_verify,
                 ),
             ),
             **kwargs,
